Move common processSlashings function to allForks
