+++ packages/beacon-state-transition/src/altair/epoch/processSlashings.ts
@@ -1,8 +1,7 @@
 import {allForks, altair} from "@chainsafe/lodestar-types";
 import {ForkName} from "@chainsafe/lodestar-params";
-
 import {CachedBeaconState, IEpochProcess} from "../../allForks/util";
-import {processSlashingsAllForks} from "../../phase0/epoch/processSlashings";
+import {processSlashingsAllForks} from "../../allForks/epoch/processSlashings";
 
 export function processSlashings(state: CachedBeaconState<altair.BeaconState>, epochProcess: IEpochProcess): void {
   processSlashingsAllForks(ForkName.altair, state as CachedBeaconState<allForks.BeaconState>, epochProcess);
